Some GX work (#1998)
* Some GX work

* GXFrameBuf work

* Import GXAttr functions and enums

* Import most of GXTev

* Import most of GXPixel

* Import GXFifo

* Import some of GXInit

* Import and work on GXTransform

* Import GXPerf

* A little work on GXTexture

+++ include/dolphin/gx/GXAttr.h
@@ -24,7 +24,7 @@ void GXSetVtxDescv(GXVtxDescList* list);
 void __GXSetVCD(void);
 void __GXCalculateVLim(void);
 void GXClearVtxDesc(void);
-void GXSetVtxAttrFmt(GXVtxFmt fmt, GXAttr attr, GXCompCnt cnt, GXCompType type, u32);
+void GXSetVtxAttrFmt(GXVtxFmt fmt, GXAttr attr, GXCompCnt cnt, GXCompType type, u8);
 void GXSetVtxAttrFmtv(GXVtxFmt fmt, GXVtxAttrFmtList* list);
 void __GXSetVAT(void);
 void GXSetArray(GXAttr attr, void* base, u8);
